requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
streamlit==1.12.0
openai==1.3.5
numpy==1.22.3
scikit-learn==1.1.3
python-dotenv==1.0.0
pydantic==1.10.13
python-multipart==0.0.6
requests==2.31.0
pandas==1.5.3


